学期选择和缓存提示样式优化
