--- components/cache-tip/cache-tip.wxss
@@ -1,6 +1,0 @@
-.cache-tip {
-  color: white;
-  font-size: 20rpx !important;
-  line-height: 100%;
-  margin-top: 10rpx;
-}
